+++ test/claper/events_test.exs
@@ -232,6 +232,16 @@ defmodule Claper.EventsTest do
       assert Events.get_managed_event!(context.alice, event.uuid) == event
     end
 
+    test "get_managed_event!/3 does not duplicate when an event has multiple leaders",
+         context do
+      event = Enum.at(context.carol_active_events, 1)
+      activity_leader_fixture(%{event: event, email: "extra-leader-1@example.com"})
+      activity_leader_fixture(%{event: event, email: "extra-leader-2@example.com"})
+
+      assert Events.get_managed_event!(context.carol, event.uuid) == event
+      assert Events.get_managed_event!(context.alice, event.uuid) == event
+    end
+
     test "get_user_event!/3 gets event by owner, raises if not", context do
       event = Enum.at(context.alice_active_events, 0)
       assert Events.get_user_event!(context.alice.id, event.uuid) == event